Rename "options" to "settings"
