+++ src/settings/index.html
@@ -3,13 +3,13 @@
     <head>
         <meta charset="UTF-8" />
         <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-        <title>Options | Progtest Themes</title>
+        <title>Settings | Progtest Themes</title>
         <link rel="stylesheet" href="../external/normalize.css" />
         <link rel="preconnect" href="https://rsms.me/" />
         <link rel="stylesheet" href="https://rsms.me/inter/inter.css" />
     </head>
     <body>
-        <script type="module" src="./options.js"></script>
+        <script type="module" src="./index.js"></script>
         <script src="../external/iconify-icon.min.js"></script>
     </body>
 </html>
